requirements.txt
pandas
playwright
textual
textual-web
croniter
PyYAML
